Remove unnecessary AsyncStorage import and token handling in Google sign-in flow
