+++ package.json
@@ -11,6 +11,7 @@
     "@react-navigation/native": "^6.1.17",
     "@react-navigation/native-stack": "^6.9.26",
     "@reduxjs/toolkit": "^2.2.7",
+    "@testing-library/react-hooks": "^8.0.1",
     "axios": "^1.6.8",
     "axios-mock-adapter": "^1.22.0",
     "expo": "51.0.8",
